refactor: move message types to shared location and add session check
Update type imports from local '../types' to shared '@/types/message' for better maintainability.
Add session check before fetching volunteer applications to prevent errors when unauthenticated.
